+++ docs/src/theme.config.js
@@ -40,6 +40,7 @@ body {
   color: rgb(220 38 38 / var(--tw-text-opacity));
 }
 
+
 .hover\:cursor-pointer:hover {
   cursor: pointer;
 }
@@ -85,7 +86,7 @@ const Head = () => {
 export const Logo = () => {
   return (
     <h1 className="flex flex-row items-baseline text-2xl font-bold">
-      <span className="tracking-tight hover:cursor-pointer">
+      <span className="dark:text-white tracking-tight hover:cursor-pointer">
         {`upload`}
         <span className="text-red-600">{`thing`}</span>
         <span className="ml-1 font-semibold">docs</span>
@@ -93,6 +94,7 @@ export const Logo = () => {
     </h1>
   );
 };
+
 
 /* eslint sort-keys: error */
 /**
@@ -102,7 +104,8 @@ const config = {
   chat: {
     link: "https://discord.gg/UCXkw6xj2K",
   },
-  darkMode: false,
+
+  darkMode: true,
   editLink: {
     component: () => null,
   },
@@ -113,9 +116,12 @@ const config = {
   head: Head,
   logo: Logo,
   nextThemes: {
-    forcedTheme: "light",
+    defaultTheme: "light",
   },
   primaryHue: 348,
+  project: {
+    link: "https://github.com/pingdotgg/uploadthing",
+  },
   useNextSeoProps() {
     return {
       additionalLinkTags: [